requirement.txt
streamlit
pandas
plotly
requests
scikit-learn
joblib
python-dotenv